Add docs, link to gh page, etc.
