+++ README.md
@@ -287,6 +287,14 @@ Check the [`wnfs/examples/`][wnfs-examples] folder for more examples.
   rs-wnfs coverage
   ```
 
+- Run benchmarks
+
+  ```bash
+  rs-wnfs bench
+  ```
+
+  You can also find a nice graph of the CI benchmarks [here][benchmarks].
+
 ## Contributing
 
 ### Pre-commit Hook
@@ -306,6 +314,7 @@ We would be happy to try to answer your question or try opening a new issue on G
 
 This project is licensed under the [Apache License 2.0](https://github.com/wnfs-wg/rs-wnfs/blob/main/LICENSE).
 
+[benchmarks]: https://wnfs-wg.github.io/rs-wnfs/dev/bench/
 [blockstore-trait]: wnfs/common/blockstore.rs#L30-L86
 [hamt-wiki]: https://en.wikipedia.org/wiki/Hash_array_mapped_trie
 [ipld-spec]: https://ipld.io/
