+++ wnfs/Cargo.toml
@@ -43,7 +43,6 @@ thiserror = "1.0"
 xxhash-rust = { version = "0.8", features = ["xxh3"] }
 
 [dev-dependencies]
-criterion = { version = "0.4", features = ["async_std"] }
 env_logger = "0.9"
 proptest = "1.0"
 rand = "0.8"
